% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_output.R
\name{get_CI}
\alias{get_CI}
\title{Calculate parameter confidence intervals}
\usage{
get_CI(fit)
}
\arguments{
\item{fit}{the output of optim i.e. the fitted model for a particular subject}
}
\description{
This function calculates parameter 95% confidence intervals from the fitted model (using the hessian supplied from optim).
}
