% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_output.R
\name{get_CItable}
\alias{get_CItable}
\title{Make parameter summary table}
\usage{
get_CItable(CIlist, param_names, free_param_index, fitted)
}
\arguments{
\item{CIlist}{a list of confidence intervals and parameter estimates obtained from fitting either the single or biphasic model to each eligible subject.}

\item{param_names}{character vector of the parameter names. This should be c("A", "delta", "B", "gamma") for the biphasic model or c("B", "gamma") for the single phase model.}

\item{free_param_index}{logical vector indicating whether the parameters A, delta, B, gamma are to be included. This should be c(TRUE, TRUE, TRUE, TRUE) for the biphasic model and c(FALSE, FALSE, TRUE, TRUE) for the single phase model.}

\item{fitted}{data frame with an 'id' column of the unique identifiers for each subject represented in CIlist. Identifiers should be ordered according to their appearance in CIlist.}
}
\description{
This function collate confidence intervals and parameter estimates from all subjects (fitted with the same model) into a nice table.
}
