% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TTS.R
\name{get_parametricTTS}
\alias{get_parametricTTS}
\title{Parametric TTS function}
\usage{
get_parametricTTS(params, rootfunction, suppression_threshold, uppertime)
}
\arguments{
\item{params}{named vector of all parameters needed to compute the suppression model, V(t)}

\item{rootfunction}{specifies which function should be used to calculate the root: biphasic or single phase.}

\item{suppression_threshold}{suppression threshold: measurements below this value will be assumed to represent viral suppression. Typically this would be the detection threshold of the assay. Default value is 20.}

\item{uppertime}{numeric value indicating the maximum time that will be considered. Default value is 365.}
}
\description{
This function computes the parametric form of the time to suppression
}
