% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_fns.R
\name{get_transformed_params}
\alias{get_transformed_params}
\title{Transform parameters}
\usage{
get_transformed_params(params, param_transform_fn)
}
\arguments{
\item{params}{vector of parameters}

\item{param_transform_fn}{vector of functions for parameter transformation}
}
\description{
This function transforms parameter estimates according to user defined functions
}
