% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_fns.R
\name{remove_vl0}
\alias{remove_vl0}
\title{Prune viral load data}
\usage{
remove_vl0(id, which_ids, subset)
}
\arguments{
\item{id}{vector of subject ids}

\item{which_ids}{vector of ids that should have the first point removed}

\item{subset}{data frame to which the function should be applied}
}
\description{
This function removes the first viral load data point for specific subjects
}
