% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_time_fixed}
\alias{simulate_time_fixed}
\title{Simulate timepoints for subjects according to fixed design.}
\usage{
simulate_time_fixed(npoints, censortime, id, index, max_datapoints)
}
\arguments{
\item{npoints}{numeric value indicating the number of observations to be sampled.}

\item{censortime}{numeric value indicating the maximum time point to include in the analysis.}

\item{id}{subject id. Can be numeric or a character.}

\item{index}{numeric identifier for each subject/model combination.}

\item{max_datapoints}{numeric value indicating the maximum number of data points collected from any subject.}
}
\description{
This function simulates observed timepoints for each subject according to a fixed sampling design.
}
