% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_vl}
\alias{simulate_vl}
\title{Simulate vl for subjects}
\usage{
simulate_vl(params, timevec, id)
}
\arguments{
\item{params}{named numeric vector of parameter values to simulate the biphasic model.}

\item{timevec}{numeric vector of observed timepoints.}

\item{id}{subject id. Can be numeric or a character.}
}
\description{
This function simulates observed vl for each subject.
}
