% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_processing.R
\name{summarize_model}
\alias{summarize_model}
\title{Summarize model output}
\usage{
summarize_model(model_output, data, stats = FALSE)
}
\arguments{
\item{model_output}{output from model fitting using ushr().}

\item{data}{dataframe of original data used for model fitting. Must include named 'id' column as a subject identifier}

\item{stats}{logical indicator: should the median and sd lifespans also be returned? Default is FALSE.}
}
\description{
This function summarizes the output of model fitting..
}
\examples{

set.seed(1234567)

simulated_data <- simulate_data(nsubjects = 20)

model_output <- ushr(data = simulated_data, detection_threshold = 10)

summarize_model(model_output, data = simulated_data)

}
