% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_output.R
\name{switch_params}
\alias{switch_params}
\title{Switch names of rate parameters}
\usage{
switch_params(biphasicCI)
}
\arguments{
\item{biphasicCI}{data frame of parameter estimates and confidence intervals for the biphasic model.}
}
\description{
This function switches the names of delta and gamma estimates if gamma > delta.
}
