% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{switch_simulated_params}
\alias{switch_simulated_params}
\title{Switch names of simulated rate parameters}
\usage{
switch_simulated_params(params)
}
\arguments{
\item{params}{matrix of parameter estimates}
}
\description{
This function switches the names of delta and gamma estimates if gamma > delta.
}
