% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_fns.R
\name{transformVL}
\alias{transformVL}
\title{Transform viral load data}
\usage{
transformVL(VL)
}
\arguments{
\item{VL}{vector of viral load data}
}
\description{
This function takes the log10 transform of viral load data & checks for NAs
}
