% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_fns_triphasic.R
\name{tri_switch_params}
\alias{tri_switch_params}
\title{Switch names of rate parameters}
\usage{
tri_switch_params(triphasicCI)
}
\arguments{
\item{triphasicCI}{data frame of parameter estimates and confidence intervals for the biphasic model.}
}
\description{
This function switches the names of delta and gamma estimates if gamma > delta.
}
