% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_processing.R
\name{filter_data}
\alias{filter_data}
\title{Prepare input data}
\usage{
filter_data(data, detection_threshold = 20, censortime = 365,
  decline_buffer = 500, n_min_single = 3, threshold_buffer = 10,
  nsuppression = 1)
}
\arguments{
\item{data}{raw data set. Must have the following column names: 'id' stating the unique identifier for each subject, 'vl' stating the viral load measurements for each subject; 'time' stating the time at which each measurement was taken.}

\item{detection_threshold}{detection threshold of the assay used to measure viral load. Measurements below this value will be assumed to represent undetectable viral load levels. Default value is 20.}

\item{censortime}{the maximum time point to inculde in the analysis. Subjects who do not suppress viral load below the detection threshold within this time will be discarded from model fitting. Default value is 365.}

\item{decline_buffer}{the maximum allowable deviation of values away from a strictly decreasing sequence in viral load. This allows for e.g. measurement noise and small fluctuations in viral load. Default value is 500.}

\item{n_min_single}{the minimum number of data points required to be included in the analysis. Defaults to 3. It is highly advised not to go below this threshold.}

\item{threshold_buffer}{numerical value indicating the range above the detection threshold which represents potential skewing of model fits. Subjects with their last two data points within this range will have the last point removed. Default value is 10.}

\item{nsuppression}{numerical value (1 or 2) indicating whether suppression is defined as having one observation below the detection threshold, or two sustained observations. Default value is 1.}
}
\description{
This function prepares the raw input data for model fitting. All default values are taken from Morris et al (2019).
}
\details{
Steps include:
1. Setting values below the detection threshold to half the detection threshhold (following standard practice).
2. Filtering out subjects who do not suppress viral load below the detection threshold by a certain time.
3. Filtering out subjects who do not have a decreasing sequence of viral load (within some buffer range).
4. Filtering out subjects who do not have enough data for model fitting.
5. Removing last data point of subjects with last two points very close to the detection threshold. This prevents skewing of the model fit
}
\examples{

set.seed(1234567)

simulated_data <- simulate_data(nsubjects = 20)

filter_data(simulated_data)

}
