% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_processing.R
\name{get_plottheme}
\alias{get_plottheme}
\title{Get plotting theme}
\usage{
get_plottheme(textsize)
}
\arguments{
\item{textsize}{numeric value for base text size. Default is 9.}
}
\description{
This function sets the plotting theme for ggplot.
}
